pyproject.toml
[project]
name = "dx-heroes"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "icecream>=2.1.5",
    "loguru>=0.7.3",
    "mypy>=1.17.0",
    "pre-commit>=4.2.0",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.4",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
